+++ tasks/task_2/kulaev_e_increasing_contrast/increasing_contrast.h
@@ -9,12 +9,12 @@
 #include <boost/algorithm/clamp.hpp>
 
 
-void Stretching(std::vector<uint8_t>* local_img, uint8_t global_min,
-                        uint8_t global_max, uint8_t new_min, uint8_t new_max);
-uint8_t generateRandomNumbers(uint8_t min, uint8_t max);
-uint8_t SeqmetionMax(std::vector<uint8_t>* vec);
-uint8_t SeqmetionMin(std::vector<uint8_t>* vec);
-void ParallelStretching(std::vector<uint8_t>* image, uint8_t new_min,
-                                                uint8_t new_max);
+void Stretching(std::vector<int>* local_img, int global_min,
+                        int global_max, int new_min, int new_max);
+int generateRandomNumbers(int min, int max);
+int SeqmetionMax(std::vector<int>* vec);
+int SeqmetionMin(std::vector<int>* vec);
+void ParallelStretching(std::vector<int>* image, int new_min,
+                                                int new_max);
 
 #endif  // TASKS_TASK_2_KULAEV_E_INCREASING_CONTRAST_INCREASING_CONTRAST_H_
