+++ tasks/task_2/kulaev_e_increasing_contrast/main.cpp
@@ -9,10 +9,10 @@ TEST(Parallel_Contrast_Enha, Test_not_elems) {
     boost::mpi::communicator comm;
     int rankProc = comm.rank();
 
-    uint8_t new_min = 0;
-    uint8_t new_max = 255;
+    int new_min = 0;
+    int new_max = 255;
 
-    std::vector<uint8_t> image{};
+    std::vector<int> image{};
     std::vector copy_image(image);
 
     ParallelStretching(&image, new_min, new_max);
@@ -26,17 +26,17 @@ TEST(Parallel_Contrast_Enha, Test_only_zero) {
     boost::mpi::communicator comm;
     int rankProc = comm.rank();
 
-    uint8_t new_min = 0;
-    uint8_t new_max = 255;
+    int new_min = 0;
+    int new_max = 255;
 
-    std::vector<uint8_t> image(100, 0);
+    std::vector<int> image(100, 0);
     std::vector copy_image(image);
 
     ParallelStretching(&image, new_min, new_max);
 
     if (rankProc == 0) {
-        uint8_t global_min = SeqmetionMin(&copy_image);
-        uint8_t global_max = SeqmetionMax(&copy_image);
+        int global_min = SeqmetionMin(&copy_image);
+        int global_max = SeqmetionMax(&copy_image);
 
         Stretching(&copy_image, global_min, global_max, new_min, new_max);
 
@@ -48,10 +48,10 @@ TEST(Parallel_Contrast_Enha, Test_rand_elems) {
     boost::mpi::communicator comm;
     int rankProc = comm.rank();
 
-    uint8_t new_min = 0;
-    uint8_t new_max = 255;
+    int new_min = 0;
+    int new_max = 255;
 
-    std::vector<uint8_t> image(625);
+    std::vector<int> image(625);
     for (auto &elem : image) elem = generateRandomNumbers(0, 255);
 
     std::vector copy_image(image);
@@ -59,8 +59,8 @@ TEST(Parallel_Contrast_Enha, Test_rand_elems) {
     ParallelStretching(&image, new_min, new_max);
 
     if (rankProc == 0) {
-        uint8_t global_min = SeqmetionMin(&copy_image);
-        uint8_t global_max = SeqmetionMax(&copy_image);
+        int global_min = SeqmetionMin(&copy_image);
+        int global_max = SeqmetionMax(&copy_image);
 
         Stretching(&copy_image, global_min, global_max, new_min, new_max);
 
@@ -72,10 +72,10 @@ TEST(Parallel_Contrast_Enha, Test_half_zero) {
     boost::mpi::communicator comm;
     int rankProc = comm.rank();
 
-    uint8_t new_min = 0;
-    uint8_t new_max = 255;
+    int new_min = 0;
+    int new_max = 255;
 
-    std::vector<uint8_t> image{0, 1, 0,
+    std::vector<int> image{0, 1, 0,
                             1, 1, 1,
                             0, 1, 0};
     std::vector copy_image(image);
@@ -83,8 +83,8 @@ TEST(Parallel_Contrast_Enha, Test_half_zero) {
     ParallelStretching(&image, new_min, new_max);
 
     if (rankProc == 0) {
-        uint8_t global_min = SeqmetionMin(&copy_image);
-        uint8_t global_max = SeqmetionMax(&copy_image);
+        int global_min = SeqmetionMin(&copy_image);
+        int global_max = SeqmetionMax(&copy_image);
 
         Stretching(&copy_image, global_min, global_max, new_min, new_max);
 
@@ -96,18 +96,18 @@ TEST(Parallel_Contrast_Enha, Test_base) {
     boost::mpi::communicator comm;
     int rankProc = comm.rank();
 
-    uint8_t new_min = 0;
-    uint8_t new_max = 255;
+    int new_min = 0;
+    int new_max = 255;
 
-    std::vector<uint8_t> image{1, 5, 5, 6, 7, 8, 8, 8,
+    std::vector<int> image{1, 5, 5, 6, 7, 8, 8, 8,
                             10, 12, 15, 16, 17, 18, 255, 0};
     std::vector copy_image(image);
 
     ParallelStretching(&image, new_min, new_max);
 
     if (rankProc == 0) {
-        uint8_t global_min = SeqmetionMin(&copy_image);
-        uint8_t global_max = SeqmetionMax(&copy_image);
+        int global_min = SeqmetionMin(&copy_image);
+        int global_max = SeqmetionMax(&copy_image);
 
         Stretching(&copy_image, global_min, global_max, new_min, new_max);
